Working towards a redesign of nodes/links to be more like ethernet.
Now, a collection of wires connecting nodes is a "net", which is part of
a subject/observer design pattern.  Upon sending data, all attached
"netports" are notified, and they store the data in their internal buffers.

The "link" code will be scrapped in favor of writing to the "netport" of
a node (which update the net, triggering a notify), and reading the
"netport" buffer.

Dedicated readers like an "IpDatagramReader" will read from buffers,
calculate the number of bytes, and return IpDatagrams.